Share type definition.
