+++ assets/src/edit-story/elements/video/index.js
@@ -19,15 +19,26 @@
  */
 import { PanelTypes } from '../../components/panels';
 import { SHARED_DEFAULT_ATTRIBUTES } from '../shared';
+import { MEDIA_DEFAULT_ATTRIBUTES, MEDIA_PANELS } from '../media';
 export { default as Display } from './display';
 export { default as Edit } from './edit';
 export { default as Frame } from './frame';
 export { default as Output } from './output';
 export { default as LayerContent } from './layer';
 export { default as LayerIcon } from './icon';
+export {
+  canFill,
+  canFlip,
+  isMaskable,
+  isMedia,
+  hasEditMode,
+  editModeGrayout,
+  resizeRules,
+} from '../media';
 
 export const defaultAttributes = {
   ...SHARED_DEFAULT_ATTRIBUTES,
+  ...MEDIA_DEFAULT_ATTRIBUTES,
   controls: false,
   loop: false,
   autoPlay: true,
@@ -36,31 +47,10 @@ export const defaultAttributes = {
     poster: null,
     videoId: 0,
   },
-  isFill: false,
-  scale: 100,
-};
-
-export const hasEditMode = true;
-
-export const isMedia = true;
-
-export const canFlip = true;
-
-export const canFill = true;
-
-export const isMaskable = true;
-
-export const resizeRules = {
-  vertical: true,
-  horizontal: true,
-  diagonal: true,
 };
 
 export const panels = [
-  PanelTypes.BACKGROUND_SIZE_POSITION,
-  PanelTypes.LAYER_STYLE,
-  PanelTypes.BACKGROUND_DISPLAY,
-  PanelTypes.SIZE_POSITION,
+  ...MEDIA_PANELS,
   PanelTypes.VIDEO_POSTER,
   PanelTypes.MASK,
 ];
